Prepare Env

+++ alegnn/modules/architecturesTime.py
@@ -239,16 +239,16 @@ class LocalGNN_DB(nn.Module):
             # Change the node number to accommodate the new order
             nodes = self.order.index(nodes)
             # If it's int, make it a list and an array
-            nodes = np.array([nodes], dtype=np.int)
+            nodes = np.array([nodes], dtype=np.int32)
             # And repeat for the number of batches
             nodes = np.tile(nodes, batchSize)
         if type(nodes) is list:
             newNodes = [self.order.index(n) for n in nodes]
-            nodes = np.array(newNodes, dtype = np.int)
+            nodes = np.array(newNodes, dtype = np.int32)
         elif type(nodes) is np.ndarray:
             newNodes = np.array([np.where(np.array(self.order) == n)[0][0] \
                                                                 for n in nodes])
-            nodes = newNodes.astype(np.int)
+            nodes = newNodes.astype(np.int32)
         # Now, nodes is an np.int np.ndarray with shape batchSize
         
         # Build the selection matrix
@@ -495,16 +495,16 @@ class GraphRecurrentNN_DB(nn.Module):
             # Change the node number to accommodate the new order
             nodes = self.order.index(nodes)
             # If it's int, make it a list and an array
-            nodes = np.array([nodes], dtype=np.int)
+            nodes = np.array([nodes], dtype=np.int32)
             # And repeat for the number of batches
             nodes = np.tile(nodes, batchSize)
         if type(nodes) is list:
             newNodes = [self.order.index(n) for n in nodes]
-            nodes = np.array(newNodes, dtype = np.int)
+            nodes = np.array(newNodes, dtype = np.int32)
         elif type(nodes) is np.ndarray:
             newNodes = np.array([np.where(np.array(self.order) == n)[0][0] \
                                                                 for n in nodes])
-            nodes = newNodes.astype(np.int)
+            nodes = newNodes.astype(np.int32)
         # Now, nodes is an np.int np.ndarray with shape batchSize
         
         # Build the selection matrix
